Fixes: Bug 32696:Email notifications is unchecked by legacy integration provisioning
Adds a new class, panopto_version_soap_client, that checks the panopto
auth service for the version number of the server being provisioned to.
In panopto data, this version is checked before provisioning, and if it
is higher than version 5, a call is made to the new
provisioncoursewithoptions method in clientdata. An option is passed to
this method specifying that the integration does not have its own
setting for setting email notifications, so the users' email
notification settings are not reset in panopto, and remain what they
were before. provision_course and provision_course_internal were updated
to set the panopto server version number before performing provisioning
operations.